+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.2] - 2024-09-18
+
+### Added
+
+- Updates the post content metadata when the content of a 'tribe_events' post has changed.
+
+### Changed
+
+- Updates the post content for an event on the import.
+
 ## [0.6.1] - 2024-09-11
 
 - Add helper function `get_sync_waiting_list()`